+++ impl/filter_lib/fo_apf.cpp
@@ -2,19 +2,17 @@
 #include <cmath>
 #include <numbers>
 
-Biquad::tp_coeffs& FO_APF::calculate_coeffs(float filter_frequency_hz, float sampling_rate_hz)
+Biquad::tp_coeffs& FO_APF::calculate_coeffs(double filter_frequency_hz, double sampling_rate_hz)
 {
     coef_t const a
-        = (std::tan(std::numbers::pi_v<float> * filter_frequency_hz / sampling_rate_hz)
-              - 1.0f)
-        / (std::tan(std::numbers::pi_v<float> * filter_frequency_hz / sampling_rate_hz)
-            + 1.0f);
+        = (std::tan(std::numbers::pi_v<double> * filter_frequency_hz / sampling_rate_hz) - 1.0)
+        / (std::tan(std::numbers::pi_v<double> * filter_frequency_hz / sampling_rate_hz) + 1.0);
 
     m_coeffs.a0 = a;
-    m_coeffs.a1 = 1.0f;
-    m_coeffs.a2 = 0.0f;
+    m_coeffs.a1 = 1.0;
+    m_coeffs.a2 = 0.0;
     m_coeffs.b1 = a;
-    m_coeffs.b2 = 0.0f;
+    m_coeffs.b2 = 0.0;
 
     return (std::ref(m_coeffs));
 }
